requirement.txt
Flask==1.1.2
torch==1.9.0
Pillow==8.3.2
Werkzeug==2.0.1

